move server into operator
