+++ services/platform/operator/server/apps/stores.go
