+++ services/platform/operator/server/system/locator.go
@@ -5,7 +5,7 @@ import (
 	"slices"
 
 	opv1 "github.com/home-cloud-io/core/services/platform/operator/api/v1"
-	k8sclient "github.com/home-cloud-io/core/services/platform/server/k8s-client"
+	k8sclient "github.com/home-cloud-io/core/services/platform/operator/server/k8s-client"
 
 	"k8s.io/apimachinery/pkg/types"
 )
