+++ services/platform/operator/server/system/os.go
@@ -15,7 +15,7 @@ import (
 
 	dv1 "github.com/home-cloud-io/core/api/platform/daemon/v1"
 	opv1 "github.com/home-cloud-io/core/services/platform/operator/api/v1"
-	k8sclient "github.com/home-cloud-io/core/services/platform/server/k8s-client"
+	k8sclient "github.com/home-cloud-io/core/services/platform/operator/server/k8s-client"
 )
 
 type (
