+++ services/platform/operator/server/utils/strings/default.go
