+++ services/platform/operator/web-client/public/favicon.ico
