+++ services/platform/operator/web-client/public/index.html
