+++ services/platform/operator/web-client/public/manifest.json
