+++ services/platform/operator/web-client/src/App.css
