+++ services/platform/operator/web-client/src/App.test.tsx
