+++ services/platform/operator/web-client/src/assets/logo-white-flat.png
