+++ services/platform/operator/web-client/src/assets/logo.png
