+++ services/platform/operator/web-client/src/index.css
