+++ services/platform/operator/web-client/src/pages/about/Logs.tsx
