+++ services/platform/operator/web-client/src/pages/store/Store.tsx
