+++ services/platform/operator/web-client/src/react-app-env.d.ts
