--- services/platform/server/main.go
@@ -1,30 +1,0 @@
-package main
-
-import (
-	"embed"
-
-	"github.com/home-cloud-io/core/services/platform/server/apps"
-	k8sclient "github.com/home-cloud-io/core/services/platform/server/k8s-client"
-	"github.com/home-cloud-io/core/services/platform/server/system"
-	"github.com/home-cloud-io/core/services/platform/server/web"
-
-	"github.com/steady-bytes/draft/pkg/chassis"
-	"github.com/steady-bytes/draft/pkg/loggers/zerolog"
-)
-
-//go:embed web-client/dist/*
-var files embed.FS
-
-func main() {
-	c := chassis.New(zerolog.New())
-	defer c.Start()
-
-	var (
-		kclient = k8sclient.NewClient(c.Logger())
-		actl    = apps.NewController(kclient)
-		sctl    = system.NewController(c.Logger(), kclient, actl)
-	)
-
-	c = c.WithClientApplication(files, "web-client/dist").
-		WithRPCHandler(web.New(c.Logger(), actl, sctl))
-}
